Exceptionhandling implemented. ServiceImpl, Controllers chenged to accomodate exceptions. User entity getUserId return type changed from long to int

+++ FlightTicketBookingSystem/src/main/java/com/capg/entities/User.java
@@ -38,7 +38,7 @@ public class User {
 		this.password = password;
 		this.contactNo = contactNo;
 	}
-	public long getUserId() {
+	public int getUserId() {
 		return userId;
 	}
 	public void setUserId(int userId) {
